use private enum singletons for formatters in reusable message objects
