+++ log4j-api/src/main/java/org/apache/logging/log4j/message/ReusableParameterizedMessage.java
@@ -18,6 +18,7 @@ package org.apache.logging.log4j.message;
 
 import java.util.Arrays;
 
+import org.apache.logging.log4j.spi.AbstractLogger;
 import org.apache.logging.log4j.util.Constants;
 import org.apache.logging.log4j.util.PerformanceSensitive;
 import org.apache.logging.log4j.util.StringBuilders;
@@ -327,7 +328,7 @@ public class ReusableParameterizedMessage implements ReusableMessage, ParameterV
 
     @Override
     public void formatTo(final StringBuilder builder) {
-        formatter.formatTo(messagePattern, getParams(), argCount, builder);
+        Formatter.INSTANCE.formatTo(messagePattern, getParams(), argCount, builder);
     }
 
     /**
@@ -358,8 +359,38 @@ public class ReusableParameterizedMessage implements ReusableMessage, ParameterV
 
     @Override
     public MessageContentFormatter getMessageContentFormatter() {
-        return formatter;
+        return Formatter.INSTANCE;
     }
 
-    private static final MessageContentFormatter formatter = new ReusableParameterizedMessageContentFormatter();
+    private enum Formatter implements MessageContentFormatter {
+        INSTANCE;
+
+        private static final ThreadLocal<int[]> localIndices = new ThreadLocal<int[]>() {
+            @Override
+            protected int[] initialValue() {
+                return new int[256];
+            }
+        };
+
+        @Override
+        public void formatTo(String formatString, Object[] parameters, int parameterCount, StringBuilder buffer) {
+            // in the event that a parameter's toString generates a log message,
+            // avoids clobbering indices that were computed from the initial call
+            // see also LOG4J2-1583
+            if (AbstractLogger.getRecursionDepth() > 1) {
+                ParameterFormatter.formatMessage(buffer, formatString, parameters, parameterCount);
+                return;
+            }
+
+            int[] indices = localIndices.get();
+            int placeholderCount = ReusableParameterizedMessage.count(formatString, indices);
+            int usedCount = Math.min(placeholderCount, parameterCount);
+
+            if (indices[0] < 0) {
+                ParameterFormatter.formatMessage(buffer, formatString, parameters, parameterCount);
+            } else {
+                ParameterFormatter.formatMessage2(buffer, formatString, parameters, usedCount, indices);
+            }
+        }
+    }
 }
